more useable code challenge interface, some styling

+++ solver-templates/code6.py
@@ -1,5 +1,5 @@
 def make_change(c):
-	return int(c/25)+int(c%25/10)+int(c%25%10/5)+int(c%25%10%5)
+	return c/25+c%25/10+c%25%10/5+c%25%10%5
 
 import sys
 if __name__ == "__main__":
